split importLimits according to dev guidelines
